requirements.txt
gradio
openai==0.28
python-dotenv